# neus-ui/src/app/services/models/create-resource-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CreateResourceDto {
  contentPath?: string;
  description?: string;
  id?: string;
  parentResourceId?: string;
  previewResourcePath?: string;
  requiredSubLevel: 'NONE' | 'BASIC' | 'ADVANCED' | 'PREMIUM';
  title?: string;
  type: 'EXAM' | 'READING_MATERIAL' | 'VIDEO' | 'BOOK' | 'LECTURE_VIDEOS' | 'LECTURE_NOTES' | 'ERMP' | 'USMLE_STEP_1' | 'USMLE_STEP_2' | 'LECTURE';
}