# neus-ui/src/app/services/models/resource-collection-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ResourceInfoDto } from '../models/resource-info-dto';
export interface ResourceCollectionDto {
  description?: string;
  requiredSubLevel?: 'NONE' | 'BASIC' | 'ADVANCED' | 'PREMIUM';
  resourceList?: Array<ResourceInfoDto>;
  title?: string;
  type?: 'EXAM' | 'READING_MATERIAL' | 'VIDEO' | 'BOOK' | 'LECTURE_VIDEOS' | 'LECTURE_NOTES' | 'ERMP' | 'USMLE_STEP_1' | 'USMLE_STEP_2' | 'LECTURE';
}