# neus-ui/src/openapi/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenApi Specification ",
        "description": "OpenApi Documentation for Alem Equb",
        "termsOfService": "term of service",
        "contact": {
            "name": "Habtamu",
            "email": "habtamuadugna.d@gmail.com"
        },
        "license": {
            "name": "Licence name ",
            "url": "http://url licence"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8088/api",
            "description": "Local ENV"
        }
    ],
    "security": [
        {
            "keycloak": []
        }
    ],
    "paths": {
...
        "/resources/{resource-id}": {
            "get": {
                "tags": [
                    "resources"
                ],
                "operationId": "getResourceById",
                "parameters": [
                    {
                        "name": "resource-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResourceDto"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "resources"
                ],
                "operationId": "updateResource",
                "parameters": [
                    {
                        "name": "resource-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateResourceDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "resources"
                ],
                "operationId": "deleteResource",
                "parameters": [
                    {
                        "name": "resource-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/resources/content/{resource-id}": {
            "put": {
                "tags": [
                    "resources"
                ],
                "operationId": "updateResourceContent",
                "parameters": [
                    {
                        "name": "resource-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "isPreview",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "required": [
                                    "file"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "resources"
                ],
                "operationId": "deleteResourceContent",
                "parameters": [
                    {
                        "name": "resource-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "url",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "isPreview",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/subscription/{sub-plan-id}": {
            "post": {
                "tags": [
                    "subscription"
                ],
                "operationId": "createSubscription",
                "parameters": [
                    {
                        "name": "sub-plan-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubscriptionDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/resources": {
            "post": {
                "tags": [
                    "resources"
                ],
                "operationId": "createResource",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "dto": {
                                        "$ref": "#/components/schemas/CreateResourceDto"
                                    },
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "previewFile": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "required": [
                                    "dto"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/subscription/{user-id}": {
            "get": {
                "tags": [
                    "subscription"
                ],
                "operationId": "getUserSubscription",
                "parameters": [
                    {
                        "name": "user-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubscriptionDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/resources/search-parent/{title}": {
            "get": {
                "tags": [
                    "resources"
                ],
                "operationId": "searchParentResourceByTitle",
                "parameters": [
                    {
                        "name": "title",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ResourceDto"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/resources/page": {
            "get": {
                "tags": [
                    "resources"
                ],
                "operationId": "getPageOfResources",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponseResourceDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/resources/detail/{resource-id}": {
            "get": {
                "tags": [
                    "resources"
                ],
                "operationId": "getResourceDetail",
                "parameters": [
                    {
                        "name": "resource-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResourceDetailDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/resources/collection/{resource-id}": {
            "get": {
                "tags": [
                    "resources"
                ],
                "operationId": "getResourceCollection",
                "parameters": [
                    {
                        "name": "resource-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResourceCollectionDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/resources/collection-list/{lecture-id}": {
            "get": {
                "tags": [
                    "resources"
                ],
                "operationId": "getListOfCollections",
                "parameters": [
                    {
                        "name": "lecture-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ResourceInfoDto"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateResourceDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "EXAM",
                            "READING_MATERIAL",
                            "VIDEO",
                            "BOOK",
                            "LECTURE_VIDEOS",
                            "LECTURE_NOTES",
                            "ERMP",
                            "USMLE_STEP_1",
                            "USMLE_STEP_2",
                            "LECTURE"
                        ]
                    },
                    "title": {
                        "type": "string",
                        "minLength": 1
                    },
                    "requiredSubLevel": {
                        "type": "string",
                        "enum": [
                            "NONE",
                            "BASIC",
                            "ADVANCED",
                            "PREMIUM"
                        ]
                    },
                    "description": {
                        "type": "string",
                        "minLength": 1
                    },
                    "contentPath": {
                        "type": "string"
                    },
                    "previewResourcePath": {
                        "type": "string"
                    },
                    "parentResourceId": {
                        "type": "string"
                    }
                },
                "required": [
                    "requiredSubLevel",
                    "type"
                ]
            },
...
            "SubscriptionDto": {
                "type": "object",
                "properties": {
                    "externalId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "level": {
                        "type": "string",
                        "enum": [
                            "NONE",
                            "BASIC",
                            "ADVANCED",
                            "PREMIUM"
                        ]
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "ResourceDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "EXAM",
                            "READING_MATERIAL",
                            "VIDEO",
                            "BOOK",
                            "LECTURE_VIDEOS",
                            "LECTURE_NOTES",
                            "ERMP",
                            "USMLE_STEP_1",
                            "USMLE_STEP_2",
                            "LECTURE"
                        ]
                    },
                    "requiredSubLevel": {
                        "type": "string",
                        "enum": [
                            "NONE",
                            "BASIC",
                            "ADVANCED",
                            "PREMIUM"
                        ]
                    },
                    "description": {
                        "type": "string"
                    },
                    "parentResourceId": {
                        "type": "string"
                    },
                    "parentResourceTitle": {
                        "type": "string"
                    },
                    "contentPath": {
                        "type": "string"
                    },
                    "previewContentPath": {
                        "type": "string"
                    }
                }
            },
            "PageResponseResourceDto": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ResourceDto"
                        }
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "numberOfElements": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "last": {
                        "type": "boolean"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "empty": {
                        "type": "boolean"
                    }
                }
            },
            "ResourceInfoDto": {
                "type": "object",
                "properties": {
                    "resourceId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "EXAM",
                            "READING_MATERIAL",
                            "VIDEO",
                            "BOOK",
                            "LECTURE_VIDEOS",
                            "LECTURE_NOTES",
                            "ERMP",
                            "USMLE_STEP_1",
                            "USMLE_STEP_2",
                            "LECTURE"
                        ]
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "requiredSubLevel": {
                        "type": "string",
                        "enum": [
                            "NONE",
                            "BASIC",
                            "ADVANCED",
                            "PREMIUM"
                        ]
                    }
                }
            },
            "ResourceDetailDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "EXAM",
                            "READING_MATERIAL",
                            "VIDEO",
                            "BOOK",
                            "LECTURE_VIDEOS",
                            "LECTURE_NOTES",
                            "ERMP",
                            "USMLE_STEP_1",
                            "USMLE_STEP_2",
                            "LECTURE"
                        ]
                    },
                    "requiredSubLevel": {
                        "type": "string",
                        "enum": [
                            "NONE",
                            "BASIC",
                            "ADVANCED",
                            "PREMIUM"
                        ]
                    },
                    "description": {
                        "type": "string"
                    },
                    "parentResourceId": {
                        "type": "string"
                    },
                    "parentResource": {
                        "type": "string"
                    },
                    "contentPath": {
                        "type": "string"
                    }
                }
            },
            "ResourceCollectionDto": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "enum": [
                            "EXAM",
                            "READING_MATERIAL",
                            "VIDEO",
                            "BOOK",
                            "LECTURE_VIDEOS",
                            "LECTURE_NOTES",
                            "ERMP",
                            "USMLE_STEP_1",
                            "USMLE_STEP_2",
                            "LECTURE"
                        ]
                    },
                    "title": {
                        "type": "string"
                    },
                    "requiredSubLevel": {
                        "type": "string",
                        "enum": [
                            "NONE",
                            "BASIC",
                            "ADVANCED",
                            "PREMIUM"
                        ]
                    },
                    "description": {
                        "type": "string"
                    },
                    "resourceList": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ResourceInfoDto"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "keycloak": {
                "type": "oauth2",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost:9090/realms/neus/protocol/openid-connect/auth",
                        "tokenUrl": "http://localhost:9090/realms/neus/protocol/openid-connect/token",
                        "scopes": {}
                    }
                }
            }
        }
    }
}